Implement stream error handling for `LaserDevice`
This updates the `nannou_laser` dependency adding support for a stream
error callback. The error callback allows for inspecting the error and
taking some action in response. See the new `streamErrorCallback` within
the diff and the associated comments for more details on the approach
taken.

In summary, the device reports TCP connection failure via the
`SuccessStatus` and continuously attempts to re-connect to the DAC until
it is found again or the user clears the device. If the DAC is found
again and the render callback is called, we have a successful connection
and so set the `SuccessStatus`' state back to `success`.